feat: resolve relative plugins/presets in suite options
